Fix python2 calls and shebangs
Explicitely call python2 instead of 'python' since on some systems (e.g.
arch linux) python3 is the default version of python. Also do this in
shebangs.

Furthermore in shebangs, use the wrapper /usr/bin/env instead
of hard-coding the interpreter path. This makes scripts more portable,
because on some systems (mostly BSDs), bash is not in /bin/ but
somewhere else in PATH.