Use validator output if it exits first, for interactive problems
Before this change, if a validator detected, say, an invalid input
format and exited, it would close its output pipe, and the submission
would proceed to read EOF and possibly crash/TLE. The output of the
submission would then be displayed to the user.

This isn't great -- it either results in confusing judgements, or
necessitates adding convoluted clean exit protocols to the problem to
deal with these cases. In the latter case, the contestant needs to add
probably-unnecessary hard-to-test code, and the description of the
protocol can bog down the problem statement (see e.g. "Go, Gopher" from
Google Code Jam 2018: https://goo.gl/ccT6t3).

This commit takes the stand that if the validator exits first with a WA
status, that should take precedence. Note that this may result in WA
statuses where the time limit is exceeded.