Correctly capitalize Mainclass, only change first letter
Python's string `capitalize` function indeed capitalizes the first
character of the string, but it also converts all other characters
in the string to lowercase.
For example, "e_CorrecT" would be `capitalize`d to "E_correct".
However, `kotlinc` only capitalizes the first letter of the class and
leaves the other characters unchanged.
For example, "e_CorrecT" will be compiled by Kotlin to "E_CorrecTKt".
The definition of the `Mainclass` variable has been updated to fix this.
(The "Kt" suffix is handled in config/languages.yaml.)