Show some output validator feedback
If the output validator put something in feedbackdir/judgemessage.txt
or feedbackdir/teammessage.txt, then show (a summmary of) that
information, preferring the former.

The summary is just a truncation of the txt-file, except if it
is recognisatble as the output from the standard output validator,
in which case it is summarised more carefully.