verifyproblem: if the language of an output validator is not from an allowed list, warn the user
Some judging environments (such as those used in Kattis) do not have all
languages available. For example, a sandbox for judging a Python 3
submission might not have Java available. If a problem uses an output
validator written in Java, and the Java runtime is not available, the
output validator would not work. So we want to warn the user that only a
subset of languages are recommended for output validation. These
recommended languages include Python 3 (because verification runs in
Python), and compiled languages like C and C++ (because they do not need
additional libraries or runtimes).