Fix adding to PYTHONPATH
Setting PYTHONPATH=newdir:$PYTHONPATH causes an empty component to
appear in $PYTHONPATH if it was originally unset. This empty component
gets treated the same as ., and this may cause erratic behavior.

(In the case when I encountered this it was because the PYTHONPATH env
var leaked into a Python submission, causing it to perform an incorrect
import.)