Update wrapper scripts and README
I could not come up with a good way to reliably make an editable install
to work in setuptools in a way that kept our support tooling available.

This reverts my earlier change of the README where I intended us to move
to using an editable install for development work, and instead just
updates our wrapper scripts to make use of a venv to pull in the correct
dependencies.