Restore old API for accessing parts of a problem. Simplify part setup
We don't really offer a stable API, but this restores most of the old
API used to access aspects of a problem (config remaining a large,
breaking change).

The rework and API changes that had been done early in the process of
supporting 2023-07-draft do not seem to be needed, so this PR reverts
much of it to simplify verifyproblem and ease porting API users (mostly
Kattis) to the new version by keeping the number of breaking changes
down.